Move GLC library, fix import in component_tools for testing
